Rename class names and folders
